+++ source/Tests/Doubles/Weight/PourRateUpdateCallbackSpy.h
@@ -8,10 +8,10 @@ namespace coffeescales::weight
 class PourRateUpdateCallbackSpy : public PourRateUpdateCallbackInterface
 {
   public:
-    void PourRateUpdate(uint32_t gramsPerSecond);
+    void PourRateUpdate(float gramsPerSecond);
 
     bool PourRateUpdateCalled = false;
-    uint32_t GramsPerSecond = 0;
+    float GramsPerSecond = 0;
 };
 
 }
